requirements.txt
configparser
numpy
requests
jsons
loggers
bs4
lxml
mysql.connector